src/requirements.txt
PyYAML==5.3.1
kubernetes==11.0.0